# Backend/openapi.yaml
...
openapi: 3.0.3
info:
  title: SocietyEase API
  description: |
    API specification for the SocietyEase backend (Flask). Authentication uses JWT Bearer tokens. Each API endpoint is mapped to its corresponding user story using the x-user-story extension for traceability between the implementation and the project requirements.

    ## User story mapping

    Every operation carries an `x-user-story` extension. The stories are:

    | ID | User story |
    |---|---|
    | **US-01** | As a **Treasurer**, I want to generate and track maintenance invoices for every flat, so that dues are billed consistently and I can see what is outstanding. |
    | **US-02** | As a **Resident**, I want to raise a complaint about a problem in my flat or the building, so that the society has a record of it and someone is accountable for fixing it. |
    | **US-03** | As a **Resident**, I want to track the status of my complaint from open to closed, so that I know whether it is being worked on without having to chase anyone. |
    | **US-04** | As a **Association manager**, I want to assign a complaint to a maintenance worker, so that the right person is responsible and the job reaches their queue. |
    | **US-05** | As a **Owner**, I want to see the invoices raised against the flat I own, so that I can keep track of what my tenant or I owe the society. |
...

    The same table is maintained in `docs/USER_STORIES.md`.
  version: 4.1.0
servers:
- url: http://localhost:5000
...
paths:
...
  /api/expenses/:
    get:
      tags:
      - Expenses
      summary: List expenses
      description: |-
        Returns every logged society expense, most recent first, optionally narrowed by the query filters below. There is no dedicated `vendor` column on Expense — a `vendor` value searches the free-text `description` field instead. **User Story — US-14, US-18**

        Access: ADMIN, SYSTEM_ADMIN or TREASURER.
      x-user-story:
      - US-14
      - US-18
      security:
      - bearerAuth: []
      parameters:
      - name: q
        in: query
        schema:
          type: string
        description: Free-text search over description.
      - name: vendor
        in: query
        schema:
          type: string
        description: Alias for `q` — there is no dedicated vendor column, so this also searches description.
      - name: category
        in: query
        schema:
          type: string
      - name: min_amount
        in: query
        schema:
          type: number
      - name: max_amount
        in: query
        schema:
          type: number
      - name: from
        in: query
        schema:
          type: string
        description: Inclusive lower bound on expense_date (YYYY-MM-DD).
      - name: to
        in: query
        schema:
          type: string
        description: Inclusive upper bound on expense_date (YYYY-MM-DD).
      responses:
        '200':
          description: Expense list, optionally filtered.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Expense'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Missing, invalid, or expired JWT bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: expenses_get_expenses
      x-required-role: ADMIN, SYSTEM_ADMIN or TREASURER
    post:
      tags:
      - Expenses
      summary: Log expense
      description: |-
        Logs a new society expense (e.g. salary, maintenance, utilities). **User Story — US-14**

        Access: ADMIN, SYSTEM_ADMIN or TREASURER.
      x-user-story: US-14
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpenseCreateRequest'
        description: JSON object. Unknown fields are ignored; blank optional values are stored as null.
      responses:
        '201':
          description: Expense logged.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expense'
        '400':
          description: category, description, amount, or expense_date is missing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing, invalid, or expired JWT bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: expenses_add_expense
      x-required-role: ADMIN, SYSTEM_ADMIN or TREASURER
  /api/expenses/{exp_id}:
    put:
      tags:
      - Expenses
      summary: Update expense
      description: |-
        Updates the description, amount, category, or receipt URL of an existing expense. **User Story — US-14**

        Access: ADMIN, SYSTEM_ADMIN or TREASURER.
      x-user-story: US-14
      security:
      - bearerAuth: []
      parameters:
      - name: exp_id
        in: path
        required: true
        schema:
          type: integer
        description: Expense id.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
                amount:
                  type: number
                category:
                  type: string
                receipt_url:
                  type: string
        description: JSON object. Unknown fields are ignored; blank optional values are stored as null.
      responses:
        '200':
          description: Expense updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expense'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Missing, invalid, or expired JWT bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: No expense with this id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: expenses_update_expense
      x-required-role: ADMIN, SYSTEM_ADMIN or TREASURER
    delete:
      tags:
      - Expenses
      summary: Delete expense
      description: |-
        Permanently deletes a logged expense. **User Story — US-14**

        Access: ADMIN, SYSTEM_ADMIN or TREASURER.
      x-user-story: US-14
      security:
      - bearerAuth: []
      parameters:
      - name: exp_id
        in: path
        required: true
        schema:
          type: integer
        description: Expense id.
      responses:
        '200':
          description: Expense deleted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Missing, invalid, or expired JWT bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: No expense with this id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: expenses_delete_expense
      x-required-role: ADMIN, SYSTEM_ADMIN or TREASURER
  /api/expenses/summary:
    get:
      tags:
      - Expenses
      summary: Monthly income/expense summary
      description: |-
        Returns total income (from paid invoices), total expense, net balance, and a per-category expense breakdown. Optionally filtered to a specific month/year via query params. **User Story — US-14**

        Access: ADMIN, SYSTEM_ADMIN or TREASURER.
      x-user-story: US-14
      security:
      - bearerAuth: []
      parameters:
      - name: month
        in: query
        schema:
          type: integer
        description: Calendar month, 1-12. Must be supplied together with `year`.
      - name: year
        in: query
        schema:
          type: integer
        description: Four-digit year, 2000-2200. Must be supplied together with `month`.
      responses:
        '200':
          description: Income/expense summary for the given (or all-time) period.
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_income:
                    type: number
                  total_expense:
                    type: number
                  net_balance:
                    type: number
                  by_category:
                    type: object
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Missing, invalid, or expired JWT bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: expenses_summary
      x-required-role: ADMIN, SYSTEM_ADMIN or TREASURER
  /api/expenses/export:
    get:
      tags:
      - Expenses
      summary: Export expenses as CSV
      description: |-
        Same query filters as GET /api/expenses/, returned as a downloadable CSV. **User Story — US-19**

        Access: ADMIN, SYSTEM_ADMIN or TREASURER.
      x-user-story: US-19
      security:
      - bearerAuth: []
      parameters:
      - name: q
        in: query
        schema:
          type: string
      - name: vendor
        in: query
        schema:
          type: string
      - name: category
        in: query
        schema:
          type: string
      - name: min_amount
        in: query
        schema:
          type: number
      - name: max_amount
        in: query
        schema:
          type: number
      - name: from
        in: query
        schema:
          type: string
      - name: to
        in: query
        schema:
          type: string
      responses:
        '200':
          description: CSV file of the filtered expense ledger.
          content:
            text/csv:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
      operationId: expenses_export
      x-required-role: ADMIN, SYSTEM_ADMIN or TREASURER
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
      - error
...
    Expense:
      type: object
      properties:
        id:
          type: integer
        category:
          type: string
        description:
          type: string
        amount:
          type: number
        expense_date:
          type: string
        paid_by:
          type: integer
        paid_by_name:
          type: string
        receipt_url:
          type: string
        created_at:
          type: string
    ExpenseCreateRequest:
      type: object
      properties:
        category:
          type: string
        description:
          type: string
        amount:
          type: number
        expense_date:
          type: string
        paid_by:
          type: integer
        receipt_url:
          type: string
      required:
      - category
      - description
      - amount
      - expense_date
...
  responses:
    BadRequest:
      description: Validation failed — a required field is missing, or a value has the wrong type/format
        (bad date, non-numeric amount, unknown enum value, or a body that is not a JSON object).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: 'category must be one of: PLUMBING, ELECTRICAL, CLEANING, SECURITY, OTHER'
    Unauthorized:
      description: 'No JWT was supplied, or it is invalid or malformed. Note: flask-jwt-extended emits
        these under a `msg` key rather than `error`.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            msg: Missing Authorization Header
    Forbidden:
      description: 'The caller is authenticated but not allowed: either the account has been deactivated,
        or the user''s role is not permitted to perform this action, or the resource belongs to another
        resident.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: You are not allowed to perform this action
    NotFound:
      description: The requested resource does not exist. Also returned when the JWT is valid but the
        user it identifies has been removed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Not found
...
    ServerError:
      description: Unexpected server error. All errors are returned as JSON; the API never returns an
        HTML error page.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Internal server error
...